Graphics fixes Changed camera res to be integer Changed fbf command to function better Added slight offset into the pixels to fix bleeding over the edge Added uniform2i to shader and compute shader Made fbf only increment when console not open Added facing command

+++ Stabby/rsrc/particles/blood.vert
@@ -3,7 +3,7 @@
 float width = 640.0;
 float height = 360.0;
 
-uniform vec2 camPos;
+uniform ivec2 camPos;
 uniform float zoom;
 
 layout(binding = 0) uniform sampler2D Image;
