+++ Stabby/rsrc/suqua/shaders/particle.vert
@@ -1,7 +1,7 @@
 #version 450 core
 
-uniform vec2 camPos = vec2(0.0, 0.0);
-uniform vec2 camRes;
+uniform vec2 camPos = vec2(0, 0);
+uniform ivec2 camRes;
 uniform vec2 zoom = vec2(1.0, 1.0);
 
 uniform vec2 windowRes;
